Adding the link creator information to the single link page
Initially I had in the top bar, but on mobile it got way too busy and made it apparent that it wasn't as important as other items that a user needs on smaller screens.

On smaller screens it stacks a the bottom at the bottom of all content, so it's there but depreciated.