Create link error message and bug fix
Changed the error message from "Invalid key" to "Enter a valid short link".

Found a bug that the destination URL allowed spaces, and then those links went to 404 pages. So now it shows an error if the user tries to create a short link with spaces in the destination url.